+++ src/ReindexerNet.Embedded/ReindexerEmbedded.Dispose.cs
@@ -5,7 +5,10 @@ namespace ReindexerNet.Embedded
 {
     public partial class ReindexerEmbedded
     {
-        private bool disposedValue; // To detect redundant calls        
+        private bool disposedValue; // To detect redundant calls      
+        /// <summary>
+        /// Query for getting namespaces.
+        /// </summary>
         protected internal const string GetNamespacesQuery = "select name FROM #namespaces";
         /// <inheritdoc/>
         protected virtual void Dispose(bool disposing)